pyproject.toml
[project]
name = "ramym-brain"
version = "0.1.0"
description = "Ramy's personal brain agent - MCP server with memory, integrations, and workflows"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0",
    "asyncpg>=0.29",
    "pgvector>=0.3",
    "sentence-transformers>=3.0",
    "httpx>=0.27",
    "msal>=1.28",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
rbrain = "src.server:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
